Working Alien AI system
Currently here's how the alien ai system works:
If it sees the player, chase the player.
If it doesn't see the player, act in the following way:
The second the player makes a sound within range of the alien (currently can only do this by jumping/running), head towards the location of the sound
If the alien has already reached the location of the last sound made/there are no sounds to make its way to, choose a location to roam to. Roaming location is determined in the following way:
If player is far enough away to be outside a certain set radius of the alien, pick a random reachable location within a given distance from the alien's current location. If the player is within the radius of the alien, it will pick a spot within a certain radius of the player. Said radius is determined by taking the current distance from the alien to the player and multiplying it by 0.3, meaning the radius around the player will be just under a third the current distance from the alien to the player (a different decimal between 0.0-1.0 may be chosen instead)